Support setting the button aspect ratio
This looks better when the aspect ratio of the grid is very different from the aspect ratio of the monitor (resulting in
very tall or very wide buttons). This is not always avoidable by using a button layout that fits the monitor (for
example, when using multiple monitors with very different aspect ratios).

The AspectBox contains both the grid and version_info, because we need to set it to expand vertically to get the buttons
to occupy the screen (this is necessary because AspectBox does not take its fixed aspect ratio into account in its
"measure" implementation, so the vertical Box it is in does not give it enough height to render the buttons at full
width). And with the AspectBox expanding vertically, if the version_info is not inside it it can end up far below the
buttons.

This mostly works, but the resulting aspect ratio only approximately matches the desired one (and it's further off than
just the label can explain, reasons unclear...).